fix: parse Skipped field as float64
TaskWarrior 3.0 serialized any numeric, even without decimal points, as
a decimal.

+++ pkg/types/task.go
@@ -12,7 +12,7 @@ type Task struct {
 	Entry       TWTime    `json:"entry"`
 	Modified    TWTime    `json:"modified"`
 	Urgency     float64   `json:"urgency,omitempty"`
-	Skipped		int		  `json:"skipped"`
+	Skipped		float64	  `json:"skipped"`
 }
 
 // type Goal struct {
